Rename SnapDialogApproval
