+++ app/components/Snaps/SnapDialogApproval/index.ts
@@ -1,3 +1,3 @@
 ///: BEGIN:ONLY_INCLUDE_IF(preinstalled-snaps)
-export { default } from './SnapsDialogApproval';
+export { default } from './SnapDialogApproval';
 ///: END:ONLY_INCLUDE_IF
